Backend/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.9.1
bcrypt==4.0.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.2.1
colorama==0.4.6
cryptography==46.0.5
Werkzeug==3.1.7 
distro==1.9.0
dnspython==2.8.0
fastapi==0.135.3
fastapi-cli==0.0.24
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.68.0
googleapis-common-protos==1.73.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.2
packaging==26.0
proto-plus==1.27.1
protobuf==6.33.6
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
passlib==1.7.4
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
sqlmodel==0.0.37
sqlparse==0.5.3
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
wheel==0.46.3
